Hard delete

+++ java/elastic-search/src/main/java/dev/sunbirdrc/elastic/ElasticServiceImpl.java
@@ -4,8 +4,6 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
-import com.jayway.jsonpath.DocumentContext;
-import com.jayway.jsonpath.JsonPath;
 import dev.sunbirdrc.pojos.ComponentHealthInfo;
 import dev.sunbirdrc.pojos.Filter;
 import dev.sunbirdrc.pojos.FilterOperators;
@@ -21,7 +19,6 @@ import org.apache.http.auth.AuthScope;
 import org.apache.http.auth.UsernamePasswordCredentials;
 import org.apache.http.client.CredentialsProvider;
 import org.apache.http.impl.client.BasicCredentialsProvider;
-import org.apache.http.impl.nio.client.HttpAsyncClientBuilder;
 import org.elasticsearch.action.admin.cluster.health.ClusterHealthRequest;
 import org.elasticsearch.action.admin.cluster.health.ClusterHealthResponse;
 import org.elasticsearch.action.admin.indices.create.CreateIndexRequest;
@@ -60,6 +57,8 @@ public class ElasticServiceImpl implements IElasticService {
     private static String userName;
     private static String password;
 
+    private static boolean isHardDeleteEnabled;
+
     public void setConnectionInfo(String connection) {
         connectionInfo = connection;
     }
@@ -72,6 +71,7 @@ public class ElasticServiceImpl implements IElasticService {
      * This method runs when the application is started in order to add all the indcies to the elastic search
      *
      * @param indices
+     *
      * @throws RuntimeException
      */
     public void init(Set<String> indices) throws RuntimeException {
@@ -223,6 +223,7 @@ public class ElasticServiceImpl implements IElasticService {
      * @param inputEntity - input json document for updating
      * @return
      */
+
     @Override
     public RestStatus updateEntity(String index, String osid, JsonNode inputEntity) {
         logger.debug("updateEntity starts with index {} and entityId {}", index, osid);
@@ -247,27 +248,19 @@ public class ElasticServiceImpl implements IElasticService {
      */
     @Override
     public RestStatus deleteEntity(String index, String osid) {
+        DeleteResponse deleteResponse = null;
         UpdateResponse response = null;
         try {
             String indexL = index.toLowerCase();
             Map<String, Object> readMap = readEntity(indexL, osid);
-            // Map<String, Object> entityMap = (Map<String, Object>) readMap.get(index);
+            if (isHardDeleteEnabled) {
+                deleteResponse = getClient(indexL).delete(new DeleteRequest(indexL, searchType, osid), RequestOptions.DEFAULT);
+                return deleteResponse.status();
+            }
             readMap.put(Constants.STATUS_KEYWORD, Constants.STATUS_INACTIVE);
             response = getClient(indexL).update(new UpdateRequest(indexL, searchType, osid).doc(readMap), RequestOptions.DEFAULT);
         } catch (NullPointerException | IOException e) {
             logger.error("exception in deleteEntity {}", e);
-            return RestStatus.NOT_FOUND;
-        }
-        return response.status();
-    }
-   // code for hard delete
-    public RestStatus hardDeleteEntity(String index, String osid) {
-        DeleteResponse response = null;
-        try {
-            String indexL = index.toLowerCase();
-            response = getClient(indexL).delete(new DeleteRequest(indexL, searchType, osid), RequestOptions.DEFAULT);
-        } catch (NullPointerException | IOException e) {
-            logger.error("exception in hardDeleteEntity {}", e);
             return RestStatus.NOT_FOUND;
         }
         return response.status();
@@ -403,5 +396,9 @@ public class ElasticServiceImpl implements IElasticService {
 
     public void setPassword(String password) {
         this.password = password;
+    }
+    public void setIsHardDeleteEnabled(boolean isHardDeleteEnabled) {
+        this.isHardDeleteEnabled = isHardDeleteEnabled;
+
     }
 }
